fix: 安全脱敏+并发泄漏+逻辑边界修复(代码审查第1-3轮)
- WEB: /api/cfg 密码/key 脱敏回传(掩码=保持); XSS 全转义(user/SSID/todo/gacha/devst)

- NET: AP 密码随机生成存NVS(去硬编码); DNS任务非阻塞自行close(fd)防泄漏; 天气缓冲4096; 待机停DNS

- SOUND: 播放状态 volatile + snd_gen uint16; I2S 初始化返回值检查

- ALARM: 去重键改本地日; SetSlot 钳位; 网页改动落地刷新

- DS1302: 读回tm_wday并把写入周期归一化; INSTRUCTION 用户名UTF-8安全截断; TIMER快进uint64; ANSWER snprintf防贴上限; GACHA utf8_len边界; main天气wbuf带余量+USER下溢+彩蛋int64

- README: 配网密码不再公开,改为随机生成

+++ components/TIMER/TIMER.c
@@ -18,7 +18,7 @@ static uint16_t tim_set_min;      /* 设定分钟 1..TIMER_MAX_MIN */
 static uint32_t tim_start;        /* 倒计时开始时刻 ms */
 static uint32_t tim_dur;          /* 倒计时总时长 ms */
 static uint32_t tim_last_sec;     /* 上次显示秒 */
-static uint32_t tim_fast_ms;      /* 快进累计补偿 ms(避免直接改 tim_start 造成下溢) */
+static uint64_t tim_fast_ms;     /* 快进累计补偿 ms(避免直接改 tim_start 造成下溢; uint64 防长按快进累积溢出) */
 static uint32_t tim_anim_start;   /* 角光标扩散动画开始时刻 */
 static int16_t  tim_corner_exp;   /* 角光标扩散量(0=收缩) */
 static int16_t  tim_slide;        /* 滑动条动画偏移(0=静止) */
@@ -37,7 +37,7 @@ static uint32_t tim_now(void)
 /* 剩余时长 ms(含快进补偿; int64 避免下溢/回绕) */
 static int64_t tim_remain(void)
 {
-    return (int64_t)tim_dur - ((int64_t)(tim_now() - tim_start) + tim_fast_ms);
+    return (int64_t)tim_dur - ((int64_t)(tim_now() - tim_start) + (int64_t)tim_fast_ms);
 }
 
 /* ---- 绘制小工具 ---- */
